Added support to return a user specific carbon footprint for a specified range
This is currently needed because the score for gamification is going to be
based on a lot more time than one week. It will also help with the range based
visualization that Shanthi is planning for the data side.

Also changed the gamification code to have the score components based on the
past year.