Define a "score" field in the user profile to maintain a stateful score that can be updated
Define the function to update it in gamified, and add a test case for it as
well. Note that the score field should really be refactored out from the user
profile since it applies only to the gamification client. The refactoring is
deferred until a bit later, when we have a better idea of what the use cases
are...