Fixes to the result template to correctly insert a URL base
The method with embedding the text as a base64 encoded string worked fine,
except for URL loading. We have a lot of relative URLs in the embedded
documents, and they were not loadable wrt the data: URL

I had initially worked around this by converting them to absolute values, but
that is not a scalable solution since it requires changing the files while
moving from dev -> production.

I now work around that by editing the included document and including a <base>
URL. I do this "the right way" by using an angular directive which decodes the
document, replaces <head> by <head><base href=""> and then generates a new
encoded value. This now works for unmodified data and game templates.