Hook up the callback to switch the default view
This is a portmanteau change, mostly focused around hooking up the default view
callback, but it has some other changes mixed in as well. A summary of the
changes follows:
- Document the reasons behind the current view selection callback (GET instead of POST)
- Implement a javascriptCallback method that can be used for such callbacks
  from our own javascript (API -> client -> choiceClient)
- Implement a generic function to store client specific state in the user profile
- Use it to store the "curr_view"
- Change the javascript to read the curr_view and use it for initialization,
  and to pass the correct query parameters to the HTTP GET call
- unit tests for the above

In addition, I pulled out the implementation of the data result into a separate
"default" client. We have discussed pulling it out as a separate client, and
this will help us get started. It will also help us invoke the method to embed
the data result into the choice screen.

Also cleaned up some of the unused functions in the javascript code. For
example, we don't really need to do anything once the iframe is loaded.