Split the storage of the results into different fields for the gamified and default cases
When I initially started tracking scores, I was lazy and stored the
carbonFootprint in the same field as the gamification score, instead of having
client specific fields. This made sense if you thought that there was one
"result" field that was populated in different ways by different clients.
Unfortunately, this model broke for the "choice" client, because we wanted to
store multiple results at the same time.

So now, I have moved the field access code into the specific clients. The
default code is in the default client. This allows the choice client to call
the gamified and default clients to do its work. This also allows us to simply
the handling of backgrond tasks, since the default background task handling is
in the default client and can be invoked just like the other clients.

I also check in a script to go through and fix the existing users. For all users
who are not in the "gamified" group, we need to move their carbon footprint
from the currentScore field to the carbon_footprint field, and delete the
currentScore and previousScore fields.