Add in support to receive data from our own data collection app
Our app sends data in the moves format, so we basically just have to invoke the
moves parsing code. I had to restructure the moves code slightly to allow this
to happen. In particular, I had to pull out the part that parsed just a trip
array, since we don't get the same unweildy segment structure as moves, and I
had to deal with the fact that we don't send a "manual" flag as part of the
input.

At this point, this works end to end on a web server running on localhost. I'm
going to push it up to production so that we can start testing out our own data
collection.