Unified the interface to the carbon module
We now always pass in the uuid, never the user directly. This allows us to
avoid confusion in what we are passing in, and subtle errors later.

Fixed callsites and test cases as well

Also added a small test case fix to do comparison with reasonable but not
excessive precision on the footprint calculations since they involve floating
point arithmetic.

All tests now pass.