Changed all the pipeline files to include actual pipelines
Several of the pipelines just included pipeline steps, but no function that
actually invoked all the functions in order.

- Changed the pipelines so that every one of them defined a runPipeline() that
  invoked the pipeline steps in order
- Changed the tests to work with the new pipelines
- Added E2E tests to every test pipeline suite that exercised the entire pipeline

Also fixed several issues while making the changes, notably the fact that the
utility model pipeline was creating the abstract superclass, AND was returning
the same class after modification.

Commented out several sections of the UserUtilityModel which need to be
commented back in after the issues are resolved.