Commit adds assert to address error that Jeff ran across (http://pastebin.com/cn0YnTje) while ground truthing clusters.
This error is the result of the the user not having any sections. I suspect that this error is the result of a typo in the user_uuid.secret file, or the result of running the truth pipeline locally without loading the necessary user data.

The assert statement gives a clearer indication of what is happening.