Integrate the canonical trip calculation with the rest of the recommender pipeline
The mode inference team had changed the canonical trip calculation to make it
work. However, the recommender pipeline was still returning the training trips
as canonical trips. This change hooks things up so that the recommender
pipeline returns the canonical trips

Extended the recommender pipeline test to test this case and fixed a bunch of
bugs found in testing. Notably, we should return an iterator, we should pass in
the Canonical_E_Mission_trip as a class to the iterator, we were assuming that
the input user UUID was encoded as a string?!

All tests currently PASS