Stupid fix to filter_modules
We didn't notice this when we were in a hurry, but the iter error was actually
the second error in the test. If we had noticed the first, it would have given
us a bigger clue.

The first is here:
Traceback (most recent call last):
  File "recommender/tripiterator.py", line 41, in __init__
    self.storedIter = query_function(user_uuid)
  File "recommender/filter_modules.py", line 76, in getCanonicalTrips
    info = get_clusters_info(uid)
  File "recommender/filter_modules.py", line 49, in get_clusters_info
    c_info.append(y, rt)
TypeError: append() takes exactly one argument (2 given)

That makes the error very obvious - we hadn't wrapped it in a tuple correctly.
After wrapping in a tuple, this is now fixed!