Add a version to the javascript to force download of the file
This is moderately bizarre, but basically, the version currently deployed to
the app store appears to aggressively cache the supporting javascript files.
That means that it doesn't find the controllers that we added later. So I'm
using a trick that I found on stackoverflow - adding a version tag forces the
javascript to be retrieved again.

With this change, the existing production version also show all the new tabs