Fix all three parts of the recommendation pipeline
The biggest change here is the restructuring of the way that the alternatives
are read. Before this, the alternatives for all trips were computed at the same
time, the trips and their corresponding alternatives zipped together, and
passed in to extract features.

However, this strategy runs into scalability issues for users with a large
number of trips. For example, we had a user (I think me) who had more than a
thousand trips. It looks like mongod has a hard limit of roughly 1020 open
connections at one time. So every time we got to this user, we would open 1020
connections simultaneously, go over the threshold, start dropping connections,
and fail any new incoming connections.

I tried working around this by increasing the number of open connections, but
that didn't seem to work, and it seemed like a flaw in the structure anyway to
have so many connections open unnecessarily.

So I refactored the code to read the alternatives for each trip as part of the
loop for the feature extraction. Note that we can go back to the zip solution
if we change the trip iterator code to open a connection to the database lazily
on demand instead of at init.

It is also not clear why the extract features code was in the individual
subclasses instead of the superclass when the code was almost identical. I
moved it to the superclass and refactored accordingly. I was also able to
refactor to handle the single trip alternative case in the recommender
pipeline.

Bunch of other fixes:
- Convert all print statements to logs and add in descriptions of what is being logged
- Handle NaN in the feature vector correctly
- Change the initialization of the feature vector from "empty" to "zero" to
  ensure that the unused lines are correctly filtered using the nonzero filter
  in the predict code
- Ensure that the crontab calls have the correct pythonpath
- Ensure that we convert the iterator of representative trips to a list so that it is not all used up. NOTE: I am not sure whether this is still required now that we don't need to zip trips any more.
- Fix the trip parsing in a couple of ways:
    - Directly use the datetime if available. Why the heck are we parsing from
      strings when we have the datetime object directly available?
    - Store the user ID
    - Fix the CanonicalTrip code since the representative trip is a section and
      not a trip. Should it be a trip? Discuss.
- Added a try/catch block in the user modelling pipeline to improve robustness

As a general note, it would be good to have the tripiterator wrap a cursor
instead of an iterator, since we want to add calls to get the length without
exhausting the iterator, and to close it, and so on.