Fix formatting + two small fixes
Most of the changes here relate to fixing formatting by replacing tabs by
spaces. This is particularly critical for python code, in which indentation is
significant - people with different tab stops will perceive different
indentations and will find it difficult to match indentation.

In addition, there are two small fixes in here.
- First, we throw a PathNotFoundException that encapsulates the OTP error
  message if OTP does not return a valid result
- Second, we change from_trip to create sections that have a "move" type,
  consistent with the new field