Top level script to evaluate various smoothing algorithms
The script has some hardcoded file paths that are defined at the top and can be
easily modified for the individual user. The script uses the set of defined
algorithms on the data and generates precision and recall values for each
section based on the user specified ground truth. It also plots the resulting
sections for visual inspection.