Implement the solution to rule them all
Expand the simple boundary and ... techniques with a port of the quality
segment technique from POSDAP and a new one that takes into account zig zags.

In order to support this, add a new function to feature calc to calculate the
speeds for a particular list instead of a section.