Tests pass (have passed three times in a row now)
Most of the issues were around imports, although I also did some more minor
renames (emission.analysis.results -> emission.analysis.result), for example.
I also had to change a bunch of file paths

Everything seems to work now. If it doesn't work, it is because it doesn't have
a test, and we should consider either writing the test or removing the
functionality.