Significantly restructure the smoothing code
The prior smoothing code had the evaluation, the smoothing algorithms and the
visualization all munged together into a giant mess. Restructure it into
separate files for the algorithms and the combination. Each algorithm type is
in a separate python file, similar to the sklearn model.

Also removed the old portmanteau file.