Couple of fixes to the section checking
Create a new method to get the points related to a section. This is a
generalization of the earlier method that passed in the section fields
separately. The older method has been restructured to call the new one.

Also added a simple call to get the section from the section id, and to add the
end of the point array as a potential split.