Refactor the location query class to work more with sections
The `get_plottable_df` method now creates a section and then calls
`get_points_for_section with the section`. This makes it easier to invoke
`get_points_for_section` with a real section object.

It also defines a enum class for the activities to make our code cleaner and a
function to return all the activities for a section.