Create a simple method to segment a trip into sections
This is purely based on the activity detection by android.

It assumes that we have all activity updates. It uses a fairly naive algorithm
with a threshold of 60% confidence, ignoring STILL, TILTING and UNKNOWN
movements, and segmenting every time the mode changes. It seems to work pretty
well for now, although I will tune it on an ongoing basis for the next few
days.