Fix the case in which there are zigzags during a stop
There are sometimes zigzags during a stop, like when I was in Millbrae.
In that case, the zigzags were detected, but the two sides were also clusters.
It is not clear if this is needed (80:20 rule), but it is relatively easy to
handle, so let us do so.

Also, add a simplification in which we return early if there are no jumps

Also, add some additional logging