Fix some minor issues with the posdap smoothing code
In particular, divide the speed by 1000 while comparing the points, because
otherwise it is too high to make any sense. This will almost certainly have to
be revisited after we fix the timestamps, but it works for now.

We should also rerun the filtering evaluation now that we know that it works!