Add support for recursive creation of wrapper classes
This assumes that every wrapper class is in a module with its name in lower
case. The convention is important for dynamic loading.

I have tested this by loading the WrapperBase, but that is not a great test because:
1) the WrapperBase doesn't have any children
2) calling any attribute directly on a WrapperBase leads to infinite recursion
at the self.props call in _getattr. Since this serves as a useful check against
the use of the abstract superclass, we choose to leave this in place.

So the test is disabled for now, and will be re-enabled when we have more real
wrapper classes that we can test against.