Add support for the trip segmentation pipeline and associated tests
Sometimes, the invocation of the trip segmentation pipeline fails (maybe
because the segmenter wants to wait to collect more data before segmenting).
This adds support for an invocation failure, which unsets the current run's
timestamp, but leaves the last run's ts unchanged. It is like the current run
never happened.

Added a test case for it.