Take a timeseries and generate trip and section decorations on top of it
This refactors the initial, default implementation of the trip segmentation to
represent one possible trip segmentation algorithm (based on dwell time). It
then puts a generic layer on top which invokes the algorithm to get the list of
stop points and creates trip and place decorations, and links them together
into a consistent timeline.

Tested against one day of my time.