Add support to the framework for sections and stops
Sections are logical parts of a trip. A new section typically occurs while
switching to a new mode, but can also be two sides of a short stop along a
route. Stops are intermediate points between sections that represent transfers
to other modes, or short stops along a route (e.g. for a bathroom break).