Change all the queries to the database to be sorted
This required passing in the sort key sometimes. If the query was on an object
that didn't have a write_ts, then the sort_key can be else. If the time query
is not specified, we use a reasonable default for the type of object (e.g.
entry: metadata.write_ts, section: end_ts, etc)