Add support for random access and filling in ends
Add support for filling in the start and end places to form a fully complete
timeline.

Also add support for random access for elements in the timeline by id. This
means that the timeline acts as a kind of in-memory cache for objects. It is
not clear whether this is a premature optimization - we need to see whether
the database accesses are just as fast.