Add our own plugin to folium to display geojson in the way that we want
The folium project has (recently?) refactored itself to support a plugin-based
architecture. This is awesome, because it makes it really easy for us to write
our own fancy javascript that displays our geojson the way that we want without
changing the base folium code.

There were some gotchas, in particular:
- creating a map with no center specified is supported in the code but breaks
  in practice
- not including the autobounds code gives really small maps that we need to zoom into

The javascript was developed by copying it into a file and working on it, and
then merging back. This was great, but lead to the most frustrating and
time-wasting experience when it turned out that I had gotten it to work by editing the
json in place, and had forgotten to make the automated change in
c5b1f09d7985e0fc3b91fc0c14929a5b2ab1a15c.

So displaying this in ipython notebook did not work and debugging it was very
frustrating because we used the minified version of the library initially.

But everything works fine now. We just need to unify the javascript functions
between the phone and here. Exactly how is TBD, probably with some angular-fu.