Add some logging statements to help with debugging + fix format of last point
Add some try/catch blocks to add clarity around errors in converting features to json

Also, the last point is not in the df, so we were adding it directly. But then
we were just adding it using the entry, which meant that the formatting change
of including the _id in the dict didn't work.