Fix a couple of serious errors in the pipeline queries
Before this change, both the get_time_range_for_usercache and
get_time_range_for_segmentation methods were invoking get_time_range_for_stage
but not returning the result. This meant that they returned None by default.

We did not catch this error immediately, since a timequery of None doesn't
filter anything, so we got all entries, and we were just testing that we got
some entries.

Returning the time queries also exposed an issue with the usercache query.
Since the usercache is not yet formatted, the write timestamps from android are
in milliseconds, not seconds. So we have to take the generic query and divide
it by 1000. But on the iOS side, we won't have to divide by 1000.

We may want to pass values in seconds from android directly in order to be
consistent. Alternatively, we can search in both milliseconds and seconds and
combine the results.