Add support for storing the recent trips from the diary into the usercache
Add support for converting trips from the last several days into geojson, and
storing it into a single cache doucment. This is consistent with the initial
plan in which there would be only one document with a particular key at a
particular time.

TODO: Decide whether we want to store this as one document per trip, or one
document for all trips in the past n days (the current option). If we want
to do the first, we can either incorporate the trip id into the key, or change
the usercache behavior.