Check in support for storing the local time in the timezone where the data was generated
This is important when we want to generate trips binned by date. When we do
this, we want to use the local time in the timezone when the data was
generated. For example, consider a trip taken by a user going to the airport to
catch a flight from the West Coast of the United States to the East Coast. Even
if the flight was in the evening West Coast time which corresponds to early
morning East Coast time, and the user is on the East Coast, she will still want
to see the trip as having occurred on the previous day since that is what
corresponds to her experience.

Similarly, an analyst in the midwest who is looking at the information to
analyse structural travel patterns will want to see the time that it occurred
in the Pacific Time Zone, not the time in the midwest or the time where the
user currently is.

We can and do, store this as formatted date/time information, but that makes it
harder to query values - for example, when retrieving the data for a particular
day.

We get around this by storing the datetime in local time as a naive datetime.

to see