Minor fixes to support populating the local dt and storing the stop as a range
The stops between sections actually represent transitions between various
modes. Because our data streams are not synchronized, we are able to identify a
transition segment, but not necessarily a point. Now, we represent the stops by
segments to be consistent with our model. We are then able to push the
complexity upstream!