Add support for a new method that returns the max timestamp for a particular field
This is used to generate the last good timestamp, so that we always spend some
time displaying something that is useful rather than blank screens. It is
unclear how useful this is once we have continuous data, but it is very useful when
we are testing against a small and ageing set of data.