Add support for storing the last processed timestamp in the database
Before this, we assumed that all data that was before the last run time had
been processed. But when data is collected on the client, it is possible for it
to have been collected before the last run but not yet transmitted to the
server. So when it does get to the server, it will be ignored.

With this change, we introduce the concept of a last_processed timestamp that
is distinct from the last run, update it for stages that use client provided
values for queries, and use it to determine the points processed in the next
run.