Read the location entries from the timeseries directly instead of through a dataframe
If we don't do this, then missing columns, like elapsedRealtimeNanos are
represented as NaN, which breaks the json format. We could replace with a
default value, or drop the offending column, but I have chosen to just not use
dataframes anymore.
https://github.com/e-mission/e-mission-server/issues/200

The reason is that with both the other options, we would need to know the set
of columns which can be missing, and would need to firefight every time a new
one changed. Since we are not actually using the dataframe features right now,
this seems more maintanable.

We can revisit if we do want to use dataframes.