Handle case in which two flags are set to true
It is possible for two flags to be set to true. According to the
CMMotionActivity documentation,

"
    "The motion-related properties of this class are not mutually exclusive.
    In other words, it is possible for more than one of the motion-related
    properties to contain the value YES. For example, if the user was driving
    in a car and the car stopped at a red light, the update event associated
    with that change in motion would have both the automotive and stationary
    properties set to YES.
"

We don't handle this case in general. But we see the case described above
fairly frequently, so we handle it for now with a type for this combination. If
we want to deal with other complex combinations, we may want to switch to a
bitmap instead.