Check in a script to debug usercache processing
Currently, if there are errors in processing the usercache, we store the
erroneous values into a separate usercache error database. But then, we want to
evaluate the errors and fix them if they reflect bugs. This is a script that
allows us to do that.

It basically copies the values back to the usercache and reruns the new move to
long term code on them. The idea is that you can rerun until you get zero
errors, or at least, zero fixable errors.