Move the "filter" field from the metadata to the data
It is unclear why it was in the metadata to begin with, given that it is only
really applicable to the location data stream. My guess is because we thought
that location data over unchanged. But we started wrapping the raw location
data on both android and iOS, so it makes sense to move the filter into the
data instead.

However, we currently have a bunch of data that was collected with the filter
in the metadata, and all incoming data is coming in also has the filter in the
metadata. So for now, we copy the filter from the metadata to the data if needed.

Once we change and update all the data collection, we can remove this.

The copying has two parts:
- a python module that can be included into the intake pipeline and copy incoming data
- a standalone script that invokes the python module for modifying existing data in the timeseries

Also added a simple test case for the module