Add in support for trip segmentation with a distance filter
At a high level, we segment the trip when we have no new entries for a time
that is greater than the time threshold. This is because we have a distance
filter, so we won't get new updates as long as we are within the filter.

In order to deal with stray noise, we extend the existing trip as long as the
new values are within the distance filter.

I've also changed the trip segmentation code to automatically use the correct
filter based on the inputs, and handled the case in which the user switches
device platforms so that we get a mixture of data from both filters.