Remove verbose logs + make it easier to ignore KeyErrors
KeyErrors are generated when the timeline is messed up. This occured when we
were processing entries that went from time -> distance -> time filters in the
same batch.

This makes it easier to ignore such errors and move on. Although in reality it
turned out that it wasn't such a smart idea, because the timeline stayed messed
up, and then even later trips were not generated properly.