Create entries if the metadata doesn't have a filter
Apparently, there were some old entries for which the metadata did not have a
filter either. All of those were from android, before we decided how we wanted
to store the values. So we manually insert "time" as the filter if none exists.

Also add a test case to check for this.