Deal with android and ios differently in the usercache code
android currently has millisecond timestamps
ios has second timestamps

we store second timestamps internally.
so for android, we need to multiply the returned query by 1000, but for ios we
don't.

So we pass in the platform and deal with it accordingly.
We also update the test case accordingly.

Note that this is part of the hack list that can be removed after all the data
is converted and the clients have been modified accordingly.