Change the android location code to retain the values after the decimal point
While converting the incoming timestamps from ms to secs.
Converting them as ints leads to a lot of identical points (zero delta) which
makes it hard to calculate the speed.

Also check in a hack to disable the assert while calculating speed. We can
re-enable it once we have processed all pending entries that we generated with
an int.