Ensure that the data returned from the usercache is sorted by the write_ts
The usercache handler assumes that data returned from the usercache is sorted
by the write_ts, because it uses the last_ts_processed as the beginning of the
next batch.

We had created the index in descending order, so all data was made available in
the wrong order.

Add sort statements to the queries to ensure that the data is returned in the
correct order. Also, if we are going to access the data in ascending order
anyway, we can change the index to also be ascending.