Handle the case in which we get spurious data
When we get spurious data, we glom it with the prior ended trip. So if the last
few points of the trip are spurious, then they are not really part of a trip, so adding a trip with them is incorrect.

So we add the new trip end only if "just_ended" is false