Add support for a new DATA_PUSHED transition on iOS
Since we push data asynchronously, and we are retaining the remote push option,
we need to know when the data push is complete so that we can return call the
completion handler. So we add a new transition to the client, which means that
we need to support the transition on the server.