Ensure that we read the locations from the new data as (lon,lat)
The current (new) format has the location coordinates as geojson, which means
that they are (lon, lat). All the new conversion code assumed that it as (lat,
lon), which resulted in common places and trips with invalid geojson - i.e.
geojson with coordinates (lat, lon).

Went through and fixed all the instances - in manual testing, the values now work.
Josh should add test cases for this since it is a frequent source of error.