Common place and trip generation code changes for new wrapper classes + remove pickling!
This currently iterates twice over the place -> successor structure as opposed to once before.

It should be fairly easy to fix - just create new entries in the database every
time you insert a new value into succcessor map. But I think that there needs
to be much more careful design of the data structures, the storage and the computation.
In particular, see https://github.com/e-mission/e-mission-server/issues/243

So we can fix this then.

Bonus fix: Remove pickling to as a technique numpy array into the database.
pickled data cannot be unpickled by javascript. We could unpickle before
storing into the usercache, but it is much more human readable to directly
store a 2-D array in mongo, which is a supported data structure.