Workaround leak in common place and common trip objects
This is a workaround for issue
https://github.com/e-mission/e-mission-server/issues/243

Since we re-create places and trips on every run, we simply delete all previous
entries before we do so.

This is clearly a band-aid over something that looks pretty annoying - create
and delete places and trips all the time. Should revisit after we have decided
how the tour model will be stored.