Handle ValueError thrown when there aren't enough unique locations for the tour model
If the cutoff is high enough, then we only get really common points, and it is
possible that the number of bins is too low for us to cluster properly. In this case,
the tour model generation code throws a ValueError. We handle this by catching
the error and skipping model creation at that time.

This is a workaround for
https://github.com/e-mission/e-mission-server/issues/244