Pass around trip objects in the new version of the cluster pipeline as well
In the old version of the cluster pipeline, we used to read the trip objects
and pass them around. In the new version of the cluster pipeline, this was
changed so that we passed around the IDs instead. This meant that we had to go
back to the database to read the trips over and over as we can see from this
pull request.

This was a particular concern while reading many trips in quick succession in
the featurization step. Since we do not have an index on the trip collection,
this caused the database to be very slow and basically stop responding.

So I went back to passing trip objects around. Now it runs fine, even for
people with a large number of trips, like me and Tom.