Return the last_processed_ts for any new configs that we just processed
And return None otherwise. This is a giant hack to address the fact that we
have a pipeline state for the output, and we need to set the "done" timestamp
for it to the actual last value of the last config that went from an
rw-document -> document.

This is a big reason to not stick with this solution. This is a hack/workaround
for now and should be removed once we come up with a final solution.

Long-term solution is tracked in:
https://github.com/e-mission/e-mission-server/issues/253