Fix usercache test to match reality
This fixes the broken build. The build was broken when I committed
https://github.com/e-mission/e-mission-server/pull/254

While we know that all tests except the last one passed, I merged without
waiting for the last set of tests to complete, since I wanted to deploy the
solution.

This broke one of the tests, but in this case, the fault was with the test and
not the code. The code passes in a list to delete_obsolete_entries, but the test
passed in the keys of a dict, which is an iterator.