Reset not just the last processed timestamp but also the current run timestamp
If we have issues with the pipeline, we need to reset it.

If the issue was such that the stage completed, but incorrectly, we need to
reset the last processed timestamp, which is what we did before this.

But if we have issues with the stage during running, the curr_run_ts is set,
and any further runs will not complete. So we need to reset the curr_run_ts as
well. That is what this change does.