Fix the delete query after moving to longTerm
It turns out that the delete query for removing entries that had been moved
into long-term was a bit broken.

Here's how the time queries played out.
If the current time at the time the pipeline is run is t_1, then:
- the pipeline state returns end_ts = t_1 - 5 (to avoid races with new data
  coming in as we are processing the current batch)
- we query for all entries < t_1 - 5
- before this, we would delete all entries < t_1 - 5, so we would delete all
  processed entries

However, after 2032751816484d77d5331eeaa10e8032c3a05864, the time queries play out as:

- the pipeline state returns end_ts = t_1 - 5 (to avoid races with new data
  coming in as we are processing the current batch)
- we query for all entries < t_1 - 5. Let us say that the last such entry is t_2.
- we delete all entries < t_2. So entries == t_2 would not be deleted and would
  be re-processed in the next round.

So we fix the query to be lte instead of lt. Now the time queries are:

- the pipeline state returns end_ts = t_1 - 5 (to avoid races with new data
  coming in as we are processing the current batch)
- we query for all entries <= t_1 - 5. Let us say that the last such entry is t_2.
- we delete all entries <= t_2. So entries == t_2 are also deleted and everything works