Move the segmentation results - trip, place, section and stop - to entries in a timeseries
Before this, we had separate collections for each of the segmentation results
- trip, place, section and stop. But that lead to a lot of repeated code for
  creation, querying and updating (CRUD methods). We now want to store cached
values for cleaned versions of these objects, in particular, for cleaned
trips and cleaned stops. Storing them in a different collection would just
lead to more duplicate code. Storing them in the same collection with a
separate key would lead to multiple special code copies that distinguish
between objects with different keys.

But we already have code that distinguishes between objects with different
keys and that it in the timeseries. By moving the segmentation results to the
timeseries as well, we have a consistent way to add new kinds of analysis
results (cleaned trips, cleaned sections, reconsituted points...), which
should make the future much easier...