Add all the javascript requirements + display a basic map
The map is based on the stamen toner tileset, which is free to use without any
obvious limits, as long as we have the proper attribution, which we do. Since
this is the main website and can be accessed even by people who don't have the
app installed or who have the app installed but may not be tracking
information, let's use layers that we are allowed to use.