Get the webapp to work end-to-end
Woot!!

I wanted to use the directives but couldn't figure out how to make them work,
so the selection components are currently in the heatmap and the controller
directly and are duplicated.

I don't like the default HTML dropdowns, so I use an action sheet wherever I
would use a select otherwise.

In general, I don't use many selects because they are more work. Right now,
people just have to enter the year, month, day, etc.

This is an ugly UI, but that's what Eric and Kris will be working on over the
summer...