Move raw trip/place/section/stop objects to the analysis TSDB
The prior version of this script copied objects from their base collection,
but to the timeseries database instead of the analysis timeseries database.
We now copy them over to the analysis TSDB.
We also stop removing entries so that we don't inadvertently lose them

Once this script has been successfully run, we should write and run a separate
script to delete the entries.

This is being tracked by
https://github.com/e-mission/e-mission-server/issues/291