Remove old, obsolete code to improve code clarity
Before this, we had a different trip format, in which the points for the trips
were stored directly in the sections. We have since switched to a format in
which the points are stored in a timeseries and the trips and sections are
generated and stored as separate objects that simply point to the original
data.

The tour modelling code had functions that would consume both kinds of data,
choosing between them by using an "old" flag. This made the code much harder
to read, since the old code was not pulled out into a separate module but was
interspersed into several functions.

This removes all the obsolete code. In particular, it removes the `old` flag,
all conditional checks on the `old` flag and all references to the old trip
format.

Now, there are no usages of the old trip in the tour model.

```
bash-3.2$ grep -r trip_old emission/analysis/modelling/tour_model/ | wc -l
       0
```