Fix import in the utility model
There is no googlemaps in the standard path but we have checked in a version
of the google maps with some minor modifications, and are already importing
it. So we can just use it instead of the failed import. I am not quite sure
how this test was working earlier.