Convert all tests to the new format as well
It turns out that the tests were not ported over when the code was, so they
were not actually testing the current code. This was exposed when the crutch
for the old code was knocked away, and the tests all started failing. Ported
over the tests to use the new data structures.

Also did some minor refactoring/cleaning to move commonly used code into a
common module so that it could be reused. The common code included setup,
cleanup and sample trip creation functions.