Handle the case where the start/end point is None
The error in b1b469b5bab6cdbe9f9c910e1fed505f3b6a409b was not caused by stale
data in the database - it was caused by the new index that we created on the
start and end locations. The indices are sparse, so having a missing value
should be fine. But in the tests, we were passing in a `None` array and using
it as the basis for the location, so we ended up with a malformed location.

Fixing that to skip setting the location when None is passed in. That should
fix the underlying issues as well.