Add a POST API method for habiticaRegister
This is an exemplar to show how to add an externally invocable POST API method
to the e-mission server.

For further explanation and examples, please consult the documentation for bottle.
http://bottlepy.org/docs/dev/index.html

@juemura, @sunil07t, @yw374cornell: this might be helpful to you as you
consider adding other methods.

This method would be invoked by sending a HTTP POST

`$http.post()`
or
`requests.post()`

with data {'username': ..., 'email': ....}

While calling this from the emission app, you should use a helper function
defined in `CommHelper` so that the JWT is inserted correctly into the HTTP
POST message.