Change the expected format slightly to match the client side API
The client side user authenticated calls assume that the payload has a single
key and data. Right now, we only need the `username` key, but it would be nice
to support other keys as well if we choose to later (e.g. for the email if
different from our email).

Also, it is a better template this way since we do want to pass in more
complex objects in the future.

We can also consider getting an array of kv objects on the client, but let's
make progress with this for now...