Implement generic habitica proxy and expose it via the API
This allows us to invoke habitica methods after mapping from our UUID to the
habitica UUID without duplicating the entire habitica API, which is pretty
large, TBH.

The new method is implemented in `proxy.py` (which is renamed from
`register.py` and has been tested with `TestHabiticaRegister.py`).

We've also exposed this proxy via an API method so that we can reuse it on the
phone as well.

The API is a POST method which expects JSON with the following format

```
    {'callOpts': {'method': ...,
                  'method_url': ....
                  'method_args': ....}
```

Note that although the proxy has been tested, the API has not been tested.
But it is fairly straightforward, so should work with minor bug fixes.