Retrieve entries corresponding to rows instead of directly wrapping them
Before this if we had a dataframe row that represented (say) a location, we
could just wrap it right back and get a location. But now that we expand the
local_date columns, this is no longer possible. Instead of recreating the
structure, we just look up the item again since we can do so directly by id.

- we add a new function in builtin_timeseries for the lookup
- we use it to lookup the start and end points for both trips and sections