Handle rollover filters correctly
As described in
https://github.com/e-mission/e-mission-server/issues/311
filtering on rollover ranges is currently broken. This is most obvious around
the start of a new month, when the default query of "past two days" straddles
the month boundary and is a rollover query.

This fixes that case.

While fixing it, I realized that there is a difference between range queries
and filter queries, and we have implemented filter queries here. This means
that the results might be somewhat unexpected, at least for rollover queries.

We defer supporting range queries until the future.